Move chain specific definitions (#440)
* moved chain specific definitions

* re-add tests for definitions

* move asset hub definition to each parachain

* move asset hub definition tests

* remove more unused constants

* remove common chain_ext file

* move pink definition

* remove all unused definitions

* remove more unused constants

* remove unused xcm asset definitions

* move testing definitions
