+++ runtime/integration-tests/src/lib.rs
@@ -13,6 +13,9 @@ mod test_macros;
 #[cfg(test)]
 mod sibling;
 
+#[cfg(test)]
+mod definitions;
+
 pub const PENDULUM_ID: u32 = 2094;
 pub const AMPLITUDE_ID: u32 = 2124;
 pub const ASSETHUB_ID: u32 = 1000;
